File Validation
File validation mostly good to go. Still a few quirks.

+++ ExternalPlugins/FileHost_External/Utils/ControlServerHandler.py
@@ -57,7 +57,7 @@ class ControlServerHandler(BaseLogging):
                 self.sync_files_action(response, headers)
             else:
                 self.logger.warning(f"Failed to retrieve files from {base_url}, code: {response.status_code}")
-                print(response.text)
+                #print(response.text)
 
 
         except Exception as e:
@@ -81,39 +81,45 @@ class ControlServerHandler(BaseLogging):
 
             # Download each file
         for file_url in file_urls_dict:
-            print(file_url)
             chunk_size          = 1024
             filesize            = file_urls_dict[file_url]["filesize"]
             filename            = file_urls_dict[file_url]["filename"]
             filepath_on_server  = file_urls_dict[file_url]["filedir"]
-            #filehash = file_urls_dict[file_url]["filehash"]
-            filehash = ""
+            filehash            = file_urls_dict[file_url]["filehash"]
+            #filehash = ""
+
+            self.logger.info(f"{self.logging_info_symbol} Downloading {filename} from {file_url} to {filepath_on_server}")
 
                 #filename = os.path.basename(file_url)
             local_file_path = os.path.join(local_directory, filename)
 
+
+            ## Need to check if file EXISTS here first... can't compare a file that doesnt exist lol
+            if os.path.exists(local_file_path):
+
                 ## Need to do hash checking. IF hashes are the same, DO NOT
                 ## redownload. Saves on processing & Network bandwidth
+                if self.file_compare(local_file = local_file_path, remote_file_hash = filehash, remote_file_name=filepath_on_server):
+                    self.logger.info(f"{self.logging_info_symbol} Exact copy of {filename} already exists. Skipping download.")
+                    continue
 
-            if self.file_compare(local_file = local_file_path, remote_file_hash = filehash, remote_file_name=filename):
-                self.logger.info(f"Exact copy of {filename} already exists. Skipping download.")
+            else:
+                # Send an HTTP GET request to the file URL and save the content to a local file
+                with open(local_file_path, 'wb') as local_file:
+                    server_response = requests.get(
+                        url=f"http://127.0.0.1:5000/filehost/{filename}",
+                        headers=headers,                        
+                        stream=True  # Set stream=True to enable streaming the content
+                    )
 
-            # Send an HTTP GET request to the file URL and save the content to a local file
-            with open(local_file_path, 'wb') as local_file:
-                server_response = requests.get(
-                    url=f"http://127.0.0.1:5000/filehost/{filename}",
-                    headers=headers,
-                    stream=True  # Set stream=True to enable streaming the content
-                )
+                    if server_response.status_code == 200:
+                        for chunk in server_response.iter_content(chunk_size=chunk_size):
+                            if chunk:
+                                local_file.write(chunk)
 
-                if server_response.status_code == 200:
-                    for chunk in server_response.iter_content(chunk_size=chunk_size):
-                        if chunk:
-                            local_file.write(chunk)
-
-                    ## Additional file checks, hash checking for file integrity?
-                else:
-                    self.logger.warning(f"{self.logging_warning_symbol} Failed to download {file_url}: {server_response.status_code}")
+                        ## Additional file checks, hash checking for file integrity?
+                    else:
+                        self.logger.warning(f"{self.logging_warning_symbol} Failed to download {file_url}: {server_response.status_code}")
 
     def file_compare(self, local_file, remote_file_hash, remote_file_name) -> bool:
         ''' 
@@ -125,8 +131,6 @@ class ControlServerHandler(BaseLogging):
         md5_hash = hashlib.md5()
 
         self.logger.debug(f"{self.logging_debug_symbol} Comparing local file: {local_file} and remote file: {remote_file_name}")
-        print(local_file)
-
         # Open the file in binary mode and read it in chunks
         with open(local_file, 'rb') as file:
             while chunk := file.read(8192):
@@ -135,12 +139,14 @@ class ControlServerHandler(BaseLogging):
         # Get the MD5 hash in hexadecimal format
         remote_md5_hex = md5_hash.hexdigest()
 
-        print("MD5 Hash:", remote_md5_hex)
+        #print("MD5 Hash:", remote_md5_hex)
+        self.logger.debug(f"{self.logging_debug_symbol} Hash of {local_file}:{remote_md5_hex}\nHash of {remote_file_name}:{remote_file_hash}")
 
 
         if remote_md5_hex == remote_file_hash:
             return True
 
+        ## Files not the same
         else:
-            self.logger.info(f"{} ")
-            return false
+            self.logger.info(f"{self.logging_info_symbol} {local_file} and {remote_file_name} are not the same. Proceeding to redownload")
+            return False
